+++ src/main/java/shop/nuribooks/books/dto/contributor/ContributorResDto.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class ContributorRes {
+public class ContributorResDto {
 	private Long id;
 	private String name;
 
-	public ContributorRes(String name) {
+	public ContributorResDto(String name) {
 		this.name = name;
 	}
