+++ src/main/java/shop/nuribooks/books/dto/contributor/role/ContributorRoleReqDto.java
@@ -1,8 +1,7 @@
-package shop.nuribooks.books.dto.contributorrole;
+package shop.nuribooks.books.dto.contributor.role;
 
 import org.hibernate.validator.constraints.Length;
 
-import jakarta.persistence.Column;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -11,10 +10,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ContributorRoleReq {
+public class ContributorRoleReqDto {
 
 	@NotNull
-	@Length(max = 50)
+	@Length(min = 1, max = 50)
 	private String name;
 }
